rename folder from erc20 to token
