+++ modules/client/src/token/internal/client/encoding.ts
