+++ modules/client/src/token/internal/utils.ts
